pyproject.toml
[tool.poetry]
name = "mental-health"
version = "0.1.0"
description = "Handle conversation between patient and AI and assess risks involved in conversation"
authors = ["Vinicius Silva <vinicius.lopes.silva89@gmail.com>"]
license = "GPLv3"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.13"
fastapi = "^0.128.0"
sqlalchemy = "^2.0.45"
asyncpg = "^0.31.0"
uvicorn = "^0.40.0"
python-dotenv = "^1.2.1"
google-genai = "^1.57.0"
types-pyyaml = "^6.0.12.20250915"


[tool.poetry.group.tests.dependencies]
pytest = "^9.0.2"
pgmock = "^1.3.7"


[tool.poetry.group.dev.dependencies]
alembic = "^1.17.2"
pre-commit = "^4.5.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
